requirements.txt
streamlit
cv2
tempfile
moviepy.editor
PIL
numpy